Change Const Byte to Const Long in run_script()
Changed the stack index from const byte to const long.

Sprite scripts will need this to use as their pointer value, then somehow
associate that with their stack.

Perhaps as `stack = &(guys.spr(getbyUID(i))->stack);`

...although `stack = &(guys.spr(GuyH::getNPCIndex(ri->guyref))->stack);`
may suffice. Not sure.

+++ src/ffscript.cpp
@@ -14624,7 +14624,7 @@ void do_combotile(const bool v)
 ///----------------------------------------------------------------------------------------------------//
 
 // Let's do this
-int run_script(const byte type, const word script, const byte i)
+int run_script(const byte type, const word script, const long i)
 {
     if(Quit==qRESET || Quit==qEXIT) // In case an earlier script hung
         return 1;
